requirements.txt
requests==2.31.0
Pillow==10.1.0
python-dateutil==2.8.2